requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
